refactor: standardize configuration file names and improve formatting in XML files
