+++ convention-plugin/src/main/kotlin/io/github/aaravmahajanofficial/internal/language/KotlinConfig.kt
@@ -13,9 +13,9 @@
  * or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package io.github.aaravmahajanofficial.internal
+package io.github.aaravmahajanofficial.internal.language
 
-import io.github.aaravmahajanofficial.constants.PluginMetadata.JAVA_VERSION
+import io.github.aaravmahajanofficial.constants.PluginMetadata
 import io.github.aaravmahajanofficial.utils.libsCatalog
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
@@ -26,7 +26,7 @@ import org.jetbrains.kotlin.gradle.dsl.KotlinJvmProjectExtension
 import org.jetbrains.kotlin.gradle.dsl.KotlinVersion
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
-public class KotlinConventionManager(
+public class KotlinConfig(
     private val project: Project,
 ) {
     private val libs = project.libsCatalog()
@@ -34,7 +34,7 @@ public class KotlinConventionManager(
     public fun configure() {
         project.plugins.withId("org.jetbrains.kotlin.jvm") {
             project.configure<KotlinJvmProjectExtension> {
-                jvmToolchain(JAVA_VERSION)
+                jvmToolchain(PluginMetadata.JAVA_VERSION)
             }
 
             val kotlinVersion =
@@ -46,9 +46,9 @@ public class KotlinConventionManager(
 
             project.tasks.withType<KotlinCompile>().configureEach { t ->
                 t.compilerOptions {
-                    languageVersion.set(KotlinVersion.fromVersion(kotlinVersion))
-                    apiVersion.set(KotlinVersion.fromVersion(kotlinVersion))
-                    jvmTarget.set(JvmTarget.fromTarget(JAVA_VERSION.toString()))
+                    languageVersion.set(KotlinVersion.Companion.fromVersion(kotlinVersion))
+                    apiVersion.set(KotlinVersion.Companion.fromVersion(kotlinVersion))
+                    jvmTarget.set(JvmTarget.Companion.fromTarget(PluginMetadata.JAVA_VERSION.toString()))
                     allWarningsAsErrors.set(true)
                     progressiveMode.set(true)
                     optIn.add("kotlin.RequiresOptIn")
