+++ convention-plugin/src/main/kotlin/io/github/aaravmahajanofficial/internal/quality/GradleVersionPluginConfig.kt
