+++ convention-plugin/src/main/resources/defaults/checkstyle/suppressions.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suppressions PUBLIC "-//Checkstyle//DTD SuppressionFilter Configuration 1.0//EN" "https://checkstyle.org/dtds/suppressions_1_0.dtd">
+<!DOCTYPE suppressions PUBLIC "-//Checkstyle//DTD SuppressionFilter Configuration 1.0//EN"
+        "https://checkstyle.org/dtds/suppressions_1_0.dtd">
 <suppressions>
-  <suppress checks=".*" files="[/\\]generated-sources[/\\]"/>
-  <suppress checks=".*" files="[/\\]generated-test-sources[/\\]"/>
+    <suppress checks=".*" files="[/\\]generated-sources[/\\]"/>
+    <suppress checks=".*" files="[/\\]generated-test-sources[/\\]"/>
 </suppressions>
