+++ convention-plugin/src/main/resources/defaults/detekt/detekt-baseline.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" ?>
 <SmellBaseline>
-  <ManuallySuppressedIssues/>
-  <CurrentIssues>
-  </CurrentIssues>
+    <ManuallySuppressedIssues/>
+    <CurrentIssues>
+    </CurrentIssues>
 </SmellBaseline>
