+++ convention-plugin/src/main/resources/defaults/spotbugs/excludesFilter.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<FindBugsFilter>
+    <!-- Add your exclude patterns here -->
+</FindBugsFilter>
